Hacky implementation of selective outbound comms
At this time the python program can request an ADC value, and the
firmware responds by writing out of hardcoded dummy array (cobs
encoded). The python program is successfully decoding this array and
displaying it to  the user. This is proof of concept ish.

The problem is that this implementation is a hack. It uses globals and
if logic, where it should be using a simple function call (the function
is called "encode_and_send()"). The problem is if the behavior is moved
from main into a function call (a 1-to-1 copy) then it stops working.
The code snippet copied doesnt make any apparent references to any other
code. The behavior should be modular, but it isn't behaving.